move FavoriteAction composable to it's own source file
